# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SnapVault API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface UploadResult {
  jobId: string;
  totalFiles: number;
  status: string;
}

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  uploading: 'uploading',
  processing: 'processing',
  awaiting_confirmation: 'awaiting_confirmation',
  zipping: 'zipping',
  ready: 'ready',
  error: 'error',
} as const;

export interface FileEntry {
  filename: string;
  originalName: string;
  category: string;
  hash: string;
  isDuplicate: boolean;
  size: number;
}

export interface CategoryGroup {
  category: string;
  count: number;
  files: FileEntry[];
}

export interface CategoryBreakdown {
  jobId: string;
  categories: CategoryGroup[];
  totalFiles: number;
  duplicateCount: number;
}

export interface Job {
  jobId: string;
  status: JobStatus;
  totalFiles: number;
  processedFiles: number;
  duplicateCount: number;
  createdAt: string;
  zipReady: boolean;
  /** @nullable */
  errorMessage?: string | null;
}

/**
 * Map of filename to overridden category (user edits)
 */
export type JobConfirmInputCategoryOverrides = {[key: string]: string};

export interface JobConfirmInput {
  /** Map of filename to overridden category (user edits) */
  categoryOverrides?: JobConfirmInputCategoryOverrides;
}

export interface ConfirmResult {
  jobId: string;
  status: string;
  message: string;
}

export interface CleanupResult {
  jobId: string;
  deleted: boolean;
  message: string;
}

export type JobStatsCategoryCounts = {[key: string]: number};

export interface JobStats {
  jobId: string;
  totalFiles: number;
  duplicateCount: number;
  categoryCounts: JobStatsCategoryCounts;
  topCategory: string;
  /** @nullable */
  zipSizeBytes?: number | null;
}