# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SnapVault API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: upload
    description: File upload operations
  - name: jobs
    description: Processing job operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /jobs/{jobId}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job status and result
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /jobs/{jobId}/categories:
    get:
      operationId: getJobCategories
      tags: [jobs]
      summary: Get categorized image preview for a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Category breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryBreakdown"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /jobs/{jobId}/confirm:
    post:
      operationId: confirmJob
      tags: [jobs]
      summary: Confirm categorization and generate ZIP
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobConfirmInput"
      responses:
        "200":
          description: ZIP generation started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfirmResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /jobs/{jobId}/download:
    get:
      operationId: downloadJobZip
      tags: [jobs]
      summary: Download the organized ZIP archive
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ZIP file download
          content:
            application/zip:
              schema:
                type: string
                format: binary
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /jobs/{jobId}/cleanup:
    delete:
      operationId: cleanupJob
      tags: [jobs]
      summary: Delete all temporary files for a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cleanup complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CleanupResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /jobs/{jobId}/stats:
    get:
      operationId: getJobStats
      tags: [jobs]
      summary: Get aggregate stats for a job (total, by category, duplicates count)
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStats"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    UploadResult:
      type: object
      properties:
        jobId:
          type: string
        totalFiles:
          type: integer
        status:
          type: string
      required:
        - jobId
        - totalFiles
        - status

    JobStatus:
      type: string
      enum:
        - uploading
        - processing
        - awaiting_confirmation
        - zipping
        - ready
        - error

    FileEntry:
      type: object
      properties:
        filename:
          type: string
        originalName:
          type: string
        category:
          type: string
        hash:
          type: string
        isDuplicate:
          type: boolean
        size:
          type: integer
      required:
        - filename
        - originalName
        - category
        - hash
        - isDuplicate
        - size

    CategoryGroup:
      type: object
      properties:
        category:
          type: string
        count:
          type: integer
        files:
          type: array
          items:
            $ref: "#/components/schemas/FileEntry"
      required:
        - category
        - count
        - files

    CategoryBreakdown:
      type: object
      properties:
        jobId:
          type: string
        categories:
          type: array
          items:
            $ref: "#/components/schemas/CategoryGroup"
        totalFiles:
          type: integer
        duplicateCount:
          type: integer
      required:
        - jobId
        - categories
        - totalFiles
        - duplicateCount

    Job:
      type: object
      properties:
        jobId:
          type: string
        status:
          $ref: "#/components/schemas/JobStatus"
        totalFiles:
          type: integer
        processedFiles:
          type: integer
        duplicateCount:
          type: integer
        createdAt:
          type: string
        zipReady:
          type: boolean
        errorMessage:
          type: ["string", "null"]
      required:
        - jobId
        - status
        - totalFiles
        - processedFiles
        - duplicateCount
        - createdAt
        - zipReady

    JobConfirmInput:
      type: object
      properties:
        categoryOverrides:
          type: object
          additionalProperties:
            type: string
          description: Map of filename to overridden category (user edits)

    ConfirmResult:
      type: object
      properties:
        jobId:
          type: string
        status:
          type: string
        message:
          type: string
      required:
        - jobId
        - status
        - message

    CleanupResult:
      type: object
      properties:
        jobId:
          type: string
        deleted:
          type: boolean
        message:
          type: string
      required:
        - jobId
        - deleted
        - message

    JobStats:
      type: object
      properties:
        jobId:
          type: string
        totalFiles:
          type: integer
        duplicateCount:
          type: integer
        categoryCounts:
          type: object
          additionalProperties:
            type: integer
        topCategory:
          type: string
        zipSizeBytes:
          type: ["integer", "null"]
      required:
        - jobId
        - totalFiles
        - duplicateCount
        - categoryCounts
        - topCategory