# lib/api-zod/src/generated/types/categoryGroup.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SnapVault API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FileEntry } from './fileEntry';

export interface CategoryGroup {
  category: string;
  count: number;
  files: FileEntry[];
}