# lib/api-zod/src/generated/types/cleanupResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SnapVault API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CleanupResult {
  jobId: string;
  deleted: boolean;
  message: string;
}