# lib/api-zod/src/generated/types/confirmResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SnapVault API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ConfirmResult {
  jobId: string;
  status: string;
  message: string;
}