# lib/api-zod/src/generated/types/fileEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SnapVault API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FileEntry {
  filename: string;
  originalName: string;
  category: string;
  hash: string;
  isDuplicate: boolean;
  size: number;
}