# lib/api-zod/src/generated/types/job.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SnapVault API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobStatus } from './jobStatus';

export interface Job {
  jobId: string;
  status: JobStatus;
  totalFiles: number;
  processedFiles: number;
  duplicateCount: number;
  createdAt: string;
  zipReady: boolean;
  /** @nullable */
  errorMessage?: string | null;
}