# lib/api-zod/src/generated/types/jobConfirmInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SnapVault API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobConfirmInputCategoryOverrides } from './jobConfirmInputCategoryOverrides';

export interface JobConfirmInput {
  /** Map of filename to overridden category (user edits) */
  categoryOverrides?: JobConfirmInputCategoryOverrides;
}