# lib/api-zod/src/generated/types/jobStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SnapVault API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobStatsCategoryCounts } from './jobStatsCategoryCounts';

export interface JobStats {
  jobId: string;
  totalFiles: number;
  duplicateCount: number;
  categoryCounts: JobStatsCategoryCounts;
  topCategory: string;
  /** @nullable */
  zipSizeBytes?: number | null;
}