# lib/api-zod/src/generated/types/jobStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SnapVault API specification
 * OpenAPI spec version: 0.1.0
 */

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  uploading: 'uploading',
  processing: 'processing',
  awaiting_confirmation: 'awaiting_confirmation',
  zipping: 'zipping',
  ready: 'ready',
  error: 'error',
} as const;