# lib/api-zod/src/generated/types/uploadResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SnapVault API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UploadResult {
  jobId: string;
  totalFiles: number;
  status: string;
}